+++ src/components/Atoms/SwitchTheme/index.tsx
@@ -2,7 +2,7 @@
 
 import React, { useEffect, useState } from "react";
 import { FiMoon, FiSun } from "react-icons/fi";
-import { useLocalStorage } from "usehooks-ts";
+
 const SwitchTheme = () => {
   //we store the theme in localStorage to preserve the state on next visit with an initial theme of dark.
   const [theme, setTheme] = useState("light");